requirements.txt
bcrypt==4.0.1
fastapi==0.104.1
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
openpyxl==3.1.2
pandas==2.0.3
paramiko==3.3.1
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.23.2
pymongo==4.6.1
colorama==0.4.6